refactor(types): split lifecycle.ts into entities/lifecycle-event.ts
